requirements.txt
fastapi
uvicorn
litellm
python-dotenv
httpx
rich
mirascope
pydantic